Fix unittests
