+++ components/brave_news/browser/test/BUILD.gn
@@ -13,6 +13,7 @@ source_set("brave_news_unit_tests") {
     "//brave/components/brave_news/browser/channels_controller_unittest.cc",
     "//brave/components/brave_news/browser/combined_feed_parsing_unittest.cc",
     "//brave/components/brave_news/browser/direct_feed_controller_unittest.cc",
+    "//brave/components/brave_news/browser/direct_feed_fetcher_unittest.cc",
     "//brave/components/brave_news/browser/feed_building_unittest.cc",
     "//brave/components/brave_news/browser/html_parsing_unittest.cc",
     "//brave/components/brave_news/browser/locales_helper_unittest.cc",
